Adding initial changes on Error Handling  - Added a new type BinderErrorConfigurer  - Added new lifecycle methods on AbstractBinder as well as dependency    on the new BinderErrorConfigurer  - Ported code from AbstractMessageChannelBinder into    AbstractMessageChannelErrorConfigurer  - Modified tests to compile against new method signatures
