+++ spring-cloud-stream-binder-test/src/test/java/org/springframework/cloud/stream/binder/MessageChannelBinderSupportTests.java
@@ -24,6 +24,7 @@ import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.Registration;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.Mockito;
 
 import org.springframework.integration.codec.kryo.KryoRegistrar;
 import org.springframework.integration.codec.kryo.PojoCodec;
@@ -266,6 +267,10 @@ public class MessageChannelBinderSupportTests {
 
 	public class TestMessageChannelBinder
 			extends AbstractBinder<MessageChannel, ConsumerProperties, ProducerProperties> {
+
+		protected TestMessageChannelBinder() {
+			super(Mockito.mock(BinderErrorConfigurer.class));
+		}
 
 		@Override
 		protected Binding<MessageChannel> doBindConsumer(String name, String group, MessageChannel channel,
