+++ spring-cloud-stream/src/test/java/org/springframework/cloud/stream/binder/BinderAwareChannelResolverTests.java
@@ -252,6 +252,11 @@ public class BinderAwareChannelResolverTests {
 				destinations.get(name).unsubscribe(directHandler);
 			}
 
+			@Override
+			public MessageChannel getTarget() {
+				return null;
+			}
+
 			public boolean isBound() {
 				return bound;
 			}
