+++ spring-cloud-stream/src/test/java/org/springframework/cloud/stream/binder/ExtendedPropertiesBinderAwareChannelResolverTests.java
@@ -202,6 +202,11 @@ public class ExtendedPropertiesBinderAwareChannelResolverTests extends BinderAwa
 			}
 
 			@Override
+			public MessageChannel getTarget() {
+				return null;
+			}
+
+			@Override
 			public void unbind() {
 				bound = false;
 				destinations.get(name).unsubscribe(directHandler);
